LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-orientation.any.worker-expected.txt


FAIL Test oriented VideoFrame from ArrayBuffer assert_equals: rotation expected (number) 90 but got (undefined) undefined
FAIL Test combinations of rotation and flip assert_equals: rotation expected (number) 0 but got (undefined) undefined
FAIL Test orientation of wrapped VideoFrame assert_equals: rotation expected (number) 90 but got (undefined) undefined
